+++ package.json
@@ -2,11 +2,11 @@
   "name": "gh-canary",
   "version": "1.0.0",
   "description": "Dashboard for github PR bot notifier",
-  "main": "server/index.js",
+  "main": "src/index.js",
   "scripts": {
-    "start": "node server/index.js",
+    "start": "node src/index.js",
     "start-debug": "cross-env DEBUG=gh-canary:* npm run start",
-    "lint": "eslint server test ",
+    "lint": "eslint src/server test",
     "test": "npm run lint && mocha test/**/*-test.js",
     "test-debug": "cross-env DEBUG=gh-canary:* npm test",
     "watch": "npm test -- -w",
