+++ src/client/public/script.js
