+++ src/index.js
@@ -1,4 +1,4 @@
-const App = require('./app');
+const App = require('./server/app');
 
 if (require.main === module) {
     const app = new App();
